Extracted Query and QueryResult models to separate modules
